+++ lib/gds_api/content_store.rb
@@ -13,13 +13,6 @@ class GdsApi::ContentStore < GdsApi::Base
     get_json(content_item_url(base_path))
   end
 
-  def incoming_links!(base_path, params = {})
-    query = query_string(params)
-    get_json!("#{endpoint}/incoming-links#{base_path}#{query}")
-  rescue GdsApi::HTTPNotFound => e
-    raise ItemNotFound.build_from(e)
-  end
-
   def content_item!(base_path)
     get_json!(content_item_url(base_path))
   rescue GdsApi::HTTPNotFound => e