+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 * Removed `GdsApi::Publisher`
 * Removed `GdsApi::ExternalLinkTracker`
-* Remove deprecated `PublishingApi` endpoints (`put_content_item` and
+* Removed deprecated `PublishingApi` endpoints (`put_content_item` and
   `put_draft_content_item`).
+* Removed `ContentStore#incoming_links!`
 
 # 37.5.1